Add files via upload
The end-of-semester evaluation is critical for both instructors and institutions. A high satisfaction rating for a professor is a testament to the teacher's abilities. It appears that beautiful instructors are more likely to earn excellent ratings. The purpose of this study is to examine the aspects that contribute to a professor's assessment score. The University of Texas in Austin collects a large sample of student instructional ratings for a set of university professors and develop six independent measures of their beauty and a variety of additional descriptors for both them and their classes. The research examines the association between professor assessment score and attractiveness rating using a two-sample mean test and a multiple linear regression model. Additionally, it employs propensity score matching and a logistic regression model to account for confounding factors. The finding is statistically significant and indicates a causal relationship between academics' assessment scores and attractiveness ratings. A professor with a higher attractiveness rating is assumed to have a better assessment rating. Additionally, this study shows that students might be biased on instructors' appearance instead of evaluating them with quality of teaching.